tests/fixtures/provider_framework_agent/requirements.txt
google-genai>=1
langgraph>=0.2
ollama>=0.4
openrouter>=1
